% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/add_ruv.R
\name{add_ruv}
\alias{add_ruv}
\title{Add residual variability to data}
\usage{
add_ruv(x, ruv)
}
\arguments{
\item{x}{data}

\item{ruv}{list with arguments prop, add}
}
\description{
Add residual variability to data
}

