% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_lbw.R
\name{calc_lbw}
\alias{calc_lbw}
\title{Calculate lean body weight
from Janmahasatian S et al. Clin Pharmacokinet. 2005;44(10):1051-65}
\usage{
calc_lbw(weight = NULL, height = NULL, sex = "male")
}
\arguments{
\item{weight}{weight in kg}

\item{height}{height in cm}

\item{sex}{either male or female}
}
\description{
Calculate lean body weight
from Janmahasatian S et al. Clin Pharmacokinet. 2005;44(10):1051-65
}

