% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_t12_effective.R
\name{calc_t12_effective}
\alias{calc_t12_effective}
\title{Calculate effective half-life}
\usage{
calc_t12_effective(t1, t2, y1, y2)
}
\arguments{
\item{t1}{first sampling timepoint}

\item{t2}{second sampling timepoint}

\item{y1}{first sample value}

\item{y2}{second sample value}
}
\description{
based on two sampling points (in same interval)
}

