% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_creat_assay.R
\name{convert_creat_assay}
\alias{convert_creat_assay}
\title{Convert serum creatinine from various assays to Jaffe}
\usage{
convert_creat_assay(scr, from = "idms", to = "jaffe")
}
\arguments{
\item{scr}{vector of serum creatinine values}

\item{from}{assay type, either `jaffe`, `enzymatic` or `idms`}

\item{to}{assay type, either `jaffe`, `enzymatic` or `idms`}
}
\description{
Convert serum creatinine from various assays to Jaffe
}

