% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_to_jaffe.R
\name{convert_to_jaffe}
\alias{convert_to_jaffe}
\title{Convert serum creatinine from various assays to Jaffe}
\usage{
convert_to_jaffe(scr, assay)
}
\arguments{
\item{scr}{vector of serum creatinine values}

\item{assay}{assay type, either `enzymatic` or `idms`}
}
\description{
Convert serum creatinine from various assays to Jaffe
}

