% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nca.R
\name{nca}
\alias{nca}
\title{Perform an NCA based on a NONMEM-style dataset}
\usage{
nca(data = NULL, dose = 100, tau = 6, method = "log_linear",
  dv_min = 0.001)
}
\arguments{
\item{data}{data.frame with time and dv columns}

\item{dose}{dose amount}

\item{tau}{dosing frequency}

\item{method}{`log_linear` or `linear`}

\item{dv_min}{minimum concentrations, lower observations will be set to this value}
}
\description{
Perform an NCA based on a NONMEM-style dataset
}

