% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_1cmt_iv_ss.R
\name{pk_1cmt_iv_ss}
\alias{pk_1cmt_iv_ss}
\title{PK of iv 1 compartment drug at steady state}
\usage{
pk_1cmt_iv_ss(t = c(0:24), dose = 100, t_inf = 1, tau = 8, CL = 5,
  V = 50, ruv = NULL)
}
\arguments{
\item{t}{vector of time}

\item{dose}{dose}

\item{t_inf}{infusion time}

\item{tau}{dosing interval}

\item{CL}{clearance}

\item{V}{volume of distribution}

\item{ruv}{residual error (list)}
}
\description{
PK of iv 1 compartment drug at steady state
}

