% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_2cmt_t12_effective.R
\name{pk_2cmt_t12_effective}
\alias{pk_2cmt_t12_effective}
\title{Calculate effective half-life for 2-compartment model}
\usage{
pk_2cmt_t12_effective(CL = 3, V = 30, Q = 2, V2 = 20, tau = 12,
  t_inf = NULL, steady_state = TRUE)
}
\arguments{
\item{CL}{clearance}

\item{V}{volume of central compartment}

\item{Q}{inter-compartimental clearance}

\item{V2}{volume of peripheral compartment}

\item{tau}{interval (hours)}

\item{t_inf}{infusion time (hours)}

\item{steady_state}{at steady state, or no?}
}
\description{
With "effective half-life" it is meant the average half-life within a specific interval
}

