% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_egfr.R
\name{calc_egfr}
\alias{calc_egfr}
\title{Calculate eGFR}
\usage{
calc_egfr(method = "cockroft_gault", sex = NULL, age = NULL, scr = NULL,
  race = "other", weight = NULL, height = NULL, bsa = NULL,
  preterm = FALSE, ckd = FALSE, bsa_method = "dubois",
  scr_unit = "mg/dl", scr_assay = "jaffe", relative = NULL,
  unit_out = "mL/min")
}
\arguments{
\item{method}{eGFR estimation method, choose from `cockroft_gault`, `mdrd`, `malmo_lund_rev`, `schwartz`}

\item{sex}{sex}

\item{age}{age}

\item{scr}{serum creatinine (mg/dL)}

\item{race}{`black` or `other`}

\item{weight}{weight}

\item{height}{height, only relevant when converting to/from BSA-relative unit}

\item{bsa}{body surface area}

\item{ckd}{chronic kidney disease? (Schwartz equation only)}

\item{bsa_method}{BSA estimation method, see `bsa()` for details}

\item{scr_unit,}{`mg/dL` or `micromol/L` (==`umol/L`)}

\item{scr_assay,}{`jaffe` or `enzymatic` or `idms`}

\item{relative}{`TRUE`/`FALSE`. Report eGFR as per 1.73 m2? Requires BSA if re-calculation required. If `NULL` (=default), will choose value typical for `method`.}

\item{unit_out}{`ml/min` (default), `L/hr`, or `mL/hr`}
}
\description{
Calculate the estimated glomerulal filtration rate, an estimate of renal function
}

