% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc_egfr.R
\name{calc_egfr}
\alias{calc_egfr}
\title{Calculate eGFR}
\usage{
calc_egfr(method = "malmo_lund_rev", sex = "male", age = 50, scr = 1,
  race = "other", weight = NULL, bsa = NULL, bsa_method = "dubois",
  scr_unit = "mg/dl", relative = FALSE, unit_out = "mL/min")
}
\arguments{
\item{method}{eGFR estimation method, choose from `cockroft_gault`, `mdrd`, `malmo_lund_rev`, `schwartz`}

\item{sex}{sex}

\item{age}{age}

\item{scr}{serum creatinine (mg/dL)}

\item{race}{`black` or `other`}

\item{weight}{weight}

\item{bsa}{body surface area}

\item{bsa_method}{BSA estimation method, see `bsa()` for details}

\item{scr_unit,}{`mg/dL` or `micromol/L` (==`umol/L`)}

\item{relative}{report eGFR as per 1.73 m2. Requires BSA depending on `method` chosen.}

\item{unit_out}{`ml/min` (default), `L/hr`, or `mL/hr`}
}
\description{
Calculate the estimated glomerulal filtration rate, an estimate of renal function
}

