% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_ffm.R
\name{calc_ffm}
\alias{calc_ffm}
\title{Calculate fat-free mass}
\usage{
calc_ffm(weight = NULL, bmi = NULL, sex = NULL)
}
\arguments{
\item{weight}{weight}

\item{bmi}{BMI}

\item{sex}{sex}
}
\description{
Get an estimate of body-surface area based on weight and height.
}
\references{
Quantification of Lean Bodyweight Sarayut Janmahasatian,1 Stephen B. Duffull,1,2 Susan Ash,3 Leigh C. Ward,4 Nuala M. Byrne5 and Bruce Green1,2
}

