% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_2cmt_t12.R
\name{pk_2cmt_t12}
\alias{pk_2cmt_t12}
\title{Calculate terminal half-life for 2-compartment model}
\usage{
pk_2cmt_t12(CL = 3, V = 30, Q = 2, V2 = 20, phase = "beta")
}
\arguments{
\item{CL}{clearance}

\item{V}{volume of central compartment}

\item{Q}{inter-compartimental clearance}

\item{V2}{volume of peripheral compartment}

\item{phase}{`alpha` or `beta` (default) to indicate initial (distribution) or terminal (elimination) phase.}
}
\description{
Calculate terminal half-life for 2-compartment model
}

